requirements.txt
# project/requirements.txt
python-dotenv
groq
requests
soundfile
https://github.com/KittenML/KittenTTS/releases/download/0.1/kittentts-0.1.0-py3-none-any.whl